Split query/evaluate.rs by phase
Decompose the 2946-line evaluator into per-phase submodules under
kernel/src/query/evaluate/: pattern matching, expression evaluation
(including Decidable QueryClass dispatch and aggregates), FIBER clause
dispatch with comorphism coercion, and RETURN/DISTINCT/ORDER-BY shaping.
The mod.rs retains only the top-level evaluate() pipeline orchestrator
and re-exports the public surface (FiberRuntime, eval_comorphism_coercion)
that external callers rely on. Tests follow the code they exercise.
